Able to edit and delete symptoms (#35)
* Add controller, model and DAO for SelectedCategory

* Save form data into DB WIP

* Fix bugs

* Add update form functinoality

* Add update functinoality in SymptomForm

* Add update test TEMP

* Log tests

* Cursor move to first

* Fix intent bundle data in main activity

* Fix table name issue

* Fix missing values

* Use date formatter

* Use symptom model rather than circle when updating

* Fix table name

* Call main activity when finish form

* Test delete operation

* Add alert when trying to delete

* Add test

* Check for deletions

* finish deletions

* Delete tests

* Get nearest symptom when press the body and add context menu to edit/delete/finish

* Add default id value for selected symptom

* Add listener to the fragment instead of body view

* Test context menu

* Context menu in fragment

* Fix context menu

* Test

* Add logic into contextMenu method

* Allow circle size logic

* Add breaks

* Add RGB logic

* Add isColorGrey

* Fix string format

* Fix typo and add new variable

* Improve method name

+++ app/src/main/java/com/project/symptoms/db/dao/SymptomCategoryDaoImpl.java
@@ -73,12 +73,12 @@ public class SymptomCategoryDaoImpl implements SymptomCategoryDao {
     }
 
     @Override
-    public boolean delete(int id) throws Exception {
+    public boolean delete(long id) throws Exception {
         return false;
     }
 
     @Override
-    public boolean update(int id, SymptomCategoryModel newValues) throws Exception {
+    public boolean update(long id, SymptomCategoryModel newValues) throws Exception {
         return false;
     }
 }
